7.7. Рефакторит редьюсер. WIP
Чем больше действий, тем не поворотливей становится редьюсер. В
демо-проекте действий немного, но редьюсер всё равно стал разрастаться.

Можно продолжать поддерживать один редьюсер, превратив его в
универсальный редьюсер, но это плохая идея. Рано или поздно в редьюсере
станет трудно ориентироваться. Лучше разбить его в соответствии
с предметной области. Таким образом получим несколько редьюсеров,
а потом объединим их с помощью `combineReducers`.

WIP: Приложение не работает, из-за рефакторинга редьюсера. Редьюсер
перестроен, нужно обновить компоненты, которые подключены к хранилищу.
