+++ src/store/game-data/game-data.js
@@ -1,0 +1,21 @@
+import {ActionType} from '../action';
+
+const initialState = {
+  questions: [],
+  isDataLoaded: false
+};
+
+const gameData = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.LOAD_QUESTIONS:
+      return {
+        ...state,
+        questions: action.payload,
+        isDataLoaded: true
+      };
+  }
+
+  return state;
+};
+
+export {gameData};
