+++ src/store/user/user.js
@@ -1,0 +1,20 @@
+import {ActionType} from '../action';
+import {AuthorizationStatus} from '../../const';
+
+const initialState = {
+  authorizationStatus: AuthorizationStatus.NO_AUTH,
+};
+
+const user = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.REQUIRED_AUTHORIZATION:
+      return {
+        ...state,
+        authorizationStatus: action.payload,
+      };
+  }
+
+  return state;
+};
+
+export {user};
